refactor: 任务监听器完成调试

+++ server/src/main/java/com/stcos/server/service/impl/FileServiceImp.java
@@ -271,16 +271,18 @@ public class FileServiceImp implements FileService {
         return sampleMetadataService.create(users);
     }
 
-    private final String PATH_ROOT = ".";
+    private final String PATH_ROOT = "./data/files";
 
     private final String PATH_FORM = "/forms";
 
     @Override
     public Resource generateFormPdf(String processId, Form form, String formName) {
         String fileName = FormUtil.formName2Chinese(formName);  // 获取表单对应的中文文件名
-        String filePathDoc = PATH_ROOT + "/" + processId + PATH_FORM + "/" + fileName + "docx";
+        String filePathDoc = PATH_ROOT + "/" + processId + PATH_FORM + "/" + fileName + ".docx";
+        String filePathPdf = PATH_ROOT + "/" + processId + PATH_FORM + "/" + fileName + ".pdf";
+        File file = new File(filePathPdf);
+        if (file.exists()) return new FileSystemResource(file);
         File docFile = FormUtil.replaceSpecialText(form, formName, filePathDoc);
-        String filePathPdf = PATH_ROOT + "/" + processId + PATH_FORM + "/" + fileName + "pdf";
         WordAndPdfUtil.word2Pdf(filePathDoc, filePathPdf);      // 将 docx 文件转换为 pdf
         //noinspection ResultOfMethodCallIgnored
         docFile.delete();                                       // 删除生成的中间 docx 文件
@@ -291,7 +293,7 @@ public class FileServiceImp implements FileService {
     @Override
     public void saveFormPdf(String processId, MultipartFile file, String formName) {
         String fileName = FormUtil.formName2Chinese(formName);     // 获取表单对应的中文文件名
-        String filePath = PATH_ROOT + "/" + processId + PATH_FORM + "/" + fileName + "pdf";
+        String filePath = PATH_ROOT + "/" + processId + PATH_FORM + "/" + fileName + ".pdf";
         try {
             FileOutputStream fOS = new FileOutputStream(filePath); // 创建文件输出流
             fOS.write(file.getBytes());                            // 将数据写入目标文件
